moved events and months to contents/programs_events
