+++ src/Components/Programs/index.js
@@ -1,4 +1,4 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 import Timeline from './Timeline';
 function Programs() {
   return (
